chore(ci_visibility): only import ddtrace when pytest ddtrace module is enabled (#8999)
This defers importing ddtrace in the `pytest`, `pytest-bdd`, and
`pytest-benchmark` plugins until after it's been determined that the
plugin is enabled with (eg: by running `pytest --ddtrace`).

The fixtures that use `ddtrace` defer importing until the feature is
actually requested due to the fact that fixtures imported later (eg:
after plugins have loaded) are not usable.

Even though this doesn't actually address the issues reported in
https://github.com/DataDog/dd-trace-py/issues/8281 or
https://github.com/DataDog/dd-trace-py/issues/4845 (due to the side
effects of `ddtrace/__init__.py`), this change is still helpful in other
ways (in particular since I'll be introducing a new version of the
`pytest` plugin shortly).

Fixtures were confirmed to still be working by:
* using `--ddtrace-patch-all`
* adding tags to spans using the `ddspan` fixture
* seeing that the `ddtracer` fixture returns the tracer object.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
