+++ ddtrace/contrib/pytest/constants.py
@@ -1,9 +1,6 @@
 FRAMEWORK = "pytest"
 KIND = "test"
 
-DDTRACE_HELP_MSG = "Enable tracing of pytest functions."
-NO_DDTRACE_HELP_MSG = "Disable tracing of pytest functions."
-DDTRACE_INCLUDE_CLASS_HELP_MSG = "Prepend 'ClassName.' to names of class-based tests."
 
 # XFail Reason
 XFAIL_REASON = "pytest.xfail.reason"
